add js fetch info to comment, new, and went over previous files to modify anything that was wrong

+++ public/js/delete.js
@@ -1,26 +1,19 @@
-async function deleteFormHandler(event) {
-    event.preventDefault();
-// separate the id from the url so we can delete from the id
-    const id = window.location.toString().split('/')[
-        window.location.toString().split('/').length - 1
-      ];
-      
-      const response = await fetch(`/api/posts/${id}`, {
-        method: 'DELETE',
-        body: JSON.stringify({
-          post_id: id
-        }),
-        headers: {
-          'Content-Type': 'application/json'
-        }
-      });
-      
-      if (response.ok) {
-        document.location.replace('/dashboard/');
-      } else {
-        alert(response.statusText);
-      }
-      
-}
+const delButtonHandler = async (event) => {
+  if (event.target.hasAttribute('post-id')) {
+    const id = event.target.getAttribute('post-id');
 
-document.querySelector('.delete-post-btn').addEventListener('click', deleteFormHandler);
+    const response = await fetch(`/api/posts/${id}`, {
+      method: 'DELETE',
+    });
+
+    if (response.ok) {
+      document.location.replace('/dashboard');
+    } else {
+      alert('Failed to delete post');
+    }
+  }
+};
+
+document
+  .querySelector('.delete-post-btn')
+  .addEventListener('click', delButtonHandler);
